+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [Unreleased]
 ### Added
 - New function `site.parseBasename`, to register new custom parsers to extract data from basenames.
-- Reload the build after change the `_config.ts` file.
+- Restart after changing the `_config.ts` or `_cms.ts` files.
 - New plugin `sri`.
 - Improved plugin docs with links to the online documentation.
 - New functions `nav.nextPage()` and `nav.previousPage()` for `nav` plugin.